Allow configuration of path to keychain file
This adds the `-keychain` parameter to all commands which can be supplied a path
to where the user want 2fa to place the keychain and read it from.

The default value of `~/.2fa` can also be overruled by setting a path in the
environment variable named `KEYCHAIN`, so it doesn't have to be supplied on
every invocation.

Priority order is:

1. The `-keychain` argument.
2. The `KEYCHAIN` environment variable.
3. Default value of `~/.2fa`.